--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,7 +2,6 @@ package hw10programoptimization
 
 import (
 	"bufio"
-	"fmt"
 	"io"
 	"regexp"
 	"strings"
@@ -23,51 +22,28 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	reader := bufio.NewReader(r)
-	i := 0
-	for {
-		line, readErr := reader.ReadBytes('\n')
-		if readErr != nil && readErr != io.EOF {
-			return result, err
-		}
-
-		var user User
-		if err := easyjson.Unmarshal(line, &user); err != nil {
-			return result, err
-		}
-		result[i] = user
-		i++
-
-		if readErr == io.EOF {
-			return result, nil
-		}
-	}
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
+	scanner := bufio.NewScanner(r)
+
+	var user User
 
 	domainRegexp, err := regexp.Compile(".+@(.+\\." + domain + ")$")
 	if err != nil {
-		return nil, err
+		return result, err
 	}
 
-	for _, user := range u {
-		found := domainRegexp.FindStringSubmatch(user.Email)
-		if len(found) > 0 {
-			fullDomain := strings.ToLower(found[1])
+	for scanner.Scan() {
+		line := scanner.Bytes()
+		if err := easyjson.Unmarshal(line, &user); err != nil {
+			return result, err
+		}
+
+		submatch := domainRegexp.FindStringSubmatch(user.Email)
+		if len(submatch) > 0 {
+			fullDomain := strings.ToLower(submatch[1])
 			result[fullDomain]++
 		}
 	}
+
 	return result, nil
 }