--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,3 +1,4 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
@@ -14,7 +15,11 @@ func TestGetDomainStat(t *testing.T) {
 {"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
-{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+{"Id":6,"Name":"Name","Username":"r","Email":"nulla@example.com.ru","Phone":"146-91-01","Password":"123","Address":"Nope"}
+{"Id":7,"Name":"Name","Username":"r","Email":"aaa@eXample.com.ru","Phone":"146-91-01","Password":"123","Address":"Nope"}
+{"Id":9,"Name":"Name","Username":"r","Email":"aaa@Example.ru","Phone":"146-91-01","Password":"123","Address":"Nope"}
+{"Id":8,"Name":"Name","Username":"r","Email":"aaa@Test.com.ru","Phone":"146-91-01","Password":"123","Address":"Nope"}`
 
 	t.Run("find 'com'", func(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
@@ -35,5 +40,17 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+
+	t.Run("find double domain 'com.ru'", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "com.ru")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{"example.com.ru": 2, "test.com.ru": 1}, result)
+	})
+
+	t.Run("find domain '.ru'", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "ru")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{"example.com.ru": 2, "test.com.ru": 1, "example.ru": 1}, result)
 	})
 }